+++ tests/test_with_dummy.py
@@ -2,12 +2,17 @@
 Tests run on a dummy package
 """
 import ubelt as ub
+from distutils.version import LooseVersion
+import sys
 from os.path import join
+from os.path import dirname
 
 
-def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
+def make_dummy_package(dpath, pkgname='mkinit_dummy_module', side_effects=0):
     """
     Creates a dummy package structure with or without __init__ files
+
+    ANY EXISTING FILES ARE DELETED
     """
     root = ub.ensuredir(join(dpath, pkgname))
     ub.delete(root)
@@ -18,10 +23,13 @@ def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
         'submod2': ub.touch(join(root, 'submod2.py')),
         'subdir1': ub.ensuredir(join(root, 'subdir1')),
         'subdir2': ub.ensuredir(join(root, 'subdir2')),
+        'longsubdir': ub.ensuredir(join(root, 'avery/long/subdir/that/goes/over80chars')),
     }
     paths['subdir1_init'] = ub.touch(join(paths['subdir1'], '__init__.py'))
     paths['subdir2_init'] = ub.touch(join(paths['subdir2'], '__init__.py'))
     paths['root_init'] = ub.touch(join(paths['root'], '__init__.py'))
+    paths['long_subdir_init'] = ub.touch(join(paths['longsubdir'], '__init__.py'))
+    paths['long_submod'] = ub.touch(join(paths['longsubdir'], 'long_submod.py'))
 
     ub.writeto(paths['subdir1_init'], ub.codeblock(
         '''
@@ -191,6 +199,52 @@ def make_dummy_package(dpath, pkgname='mkinit_dummy_module'):
         if __name__ == 'something_else':
             bad_something_else = None
         '''))
+
+    # Ensure that each submodule has an __init__
+    # (do we need this with PEP 420 anymore?)
+    root = paths['root']
+    from os.path import relpath, exists, dirname
+    import os
+
+    dpath = dirname(root)
+    init_fpaths = set()
+    for key, path in paths.items():
+        relative = relpath(path, dpath)
+        suffix = []
+        parts = relative.split(os.sep)
+        for part in parts:
+            if '.' not in part:
+                suffix.append(part)
+                middir = join(dpath, os.sep.join(suffix))
+                fpath = join(middir, '__init__.py')
+                init_fpaths.add(fpath)
+
+    for fpath in init_fpaths:
+        if not exists(fpath):
+            ub.touch(fpath)
+
+    ub.writeto(paths['long_submod'], ub.codeblock(
+        '''
+        def a_very_nested_function():
+            print('a_very_nested_function in {}'.format(__file__))
+        def another_func1():
+            pass
+        def another_func2():
+            pass
+        def another_func3():
+            pass
+        def another_func4():
+            pass
+        '''
+    ))
+
+    if side_effects:
+        with open(paths['long_submod'], 'a') as file:
+            file.write(ub.codeblock(
+                '''
+                print('NESTED SIDE EFFECT')
+                '''
+            ))
     return paths
 
 
@@ -216,8 +270,13 @@ def test_static_import_without_init():
     paths = make_dummy_package(cache_dpath)
     ub.delete(paths['root_init'])
 
+    text = mkinit.static_init(paths['long_subdir_init'])
+    print(text)
+    # check_dummy_root_init(text)
+
     modpath = paths['root']
     text = mkinit.static_init(modpath)
+    print(text)
     check_dummy_root_init(text)
 
 
@@ -260,6 +319,78 @@ def test_dynamic_init():
     for i in range(1, 15):
         want = 'good_attr_{:02d}'.format(i)
         assert want in text, 'missing {}'.format(want)
+
+
+def test_lazy_import():
+    """
+    python ~/code/mkinit/tests/test_with_dummy.py test_lazy_import
+    """
+    import pytest
+    import mkinit
+
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath)
+
+    pkg_path = paths['root']
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 1}, dry=False, recursive=True)
+
+    if LooseVersion('{}.{}'.format(*sys.version_info[0:2])) < LooseVersion('3.7'):
+        pytest.skip()
+
+    dpath = dirname(paths['root'])
+    with ub.util_import.PythonPathContext(dpath):
+        import mkinit_dummy_module
+        print('mkinit_dummy_module = {!r}'.format(mkinit_dummy_module))
+        print(dir(mkinit_dummy_module))
+        print('mkinit_dummy_module.a_very_nested_function = {!r}'.format(mkinit_dummy_module.a_very_nested_function))
+        mkinit_dummy_module.a_very_nested_function()
+
+
+def test_recursive_lazy_autogen():
+    """
+    xdoctest ~/code/mkinit/tests/test_with_dummy.py test_recursive_lazy_autogen
+    """
+    import pytest
+    import mkinit
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath, pkgname='mkinit_rec_lazy_autogen', side_effects=True)
+    pkg_path = paths['root']
+
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 1}, dry=False, recursive=True)
+
+    if LooseVersion('{}.{}'.format(*sys.version_info[0:2])) < LooseVersion('3.7'):
+        pytest.skip()
+
+    with ub.util_import.PythonPathContext(cache_dpath):
+        import mkinit_rec_lazy_autogen
+        print('mkinit_rec_lazy_autogen = {!r}'.format(mkinit_rec_lazy_autogen))
+        print('mkinit_rec_lazy_autogen.good_attr_01 = {!r}'.format(mkinit_rec_lazy_autogen.good_attr_01))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        print('mkinit_rec_lazy_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_lazy_autogen.a_very_nested_function))
+        mkinit_rec_lazy_autogen.a_very_nested_function()
+
+
+def test_recursive_eager_autogen():
+    """
+    xdoctest ~/code/mkinit/tests/test_with_dummy.py test_recursive_eager_autogen
+    """
+    import mkinit
+    cache_dpath = ub.ensure_app_cache_dir('mkinit/tests')
+    paths = make_dummy_package(cache_dpath, pkgname='mkinit_rec_eager_autogen', side_effects=True)
+    pkg_path = paths['root']
+
+    mkinit.autogen_init(pkg_path, options={'lazy_import': 0}, dry=False, recursive=True)
+
+    with ub.util_import.PythonPathContext(cache_dpath):
+        import mkinit_rec_eager_autogen
+        print('mkinit_rec_eager_autogen = {!r}'.format(mkinit_rec_eager_autogen))
+        print('mkinit_rec_eager_autogen = {!r}'.format(mkinit_rec_eager_autogen))
+        print('mkinit_rec_eager_autogen.good_attr_01 = {!r}'.format(mkinit_rec_eager_autogen.good_attr_01))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        print('mkinit_rec_eager_autogen.a_very_nested_function = {!r}'.format(mkinit_rec_eager_autogen.a_very_nested_function))
+        mkinit_rec_eager_autogen.a_very_nested_function()
 
 
 if __name__ == '__main__':